# [Inductor][CPP] Optimize WOQ INT8 wgt dequant in AMX GEMM template