@leslie-fang-intel 2024-09-26
> @leslie-fang-intel, could there have been some copy-paste error pertaining to the perf data on `M=4096, N=11008, K=4096` and `M=4096, N=4096, K=11008`? I'm using an Intel Xeon SP Gen 4 (Platinum 8468H), and the `_weight_int8pack_mm` ATen kernel doesn't have similar performance for both configs with 48 physical cores (the difference is ~6%). At your end, though, both input shapes exhibited similar performance.

I guess it may due to the system difference.